refactor: decompose Universe Builder controller and tab components (#2374)
Extract RenderTab, CompositeSheetsEditor, BucketChipStrip components,
useRenderJobQueue hook, and universeBuilderCounts/Shared lib helpers
out of the monolithic UniverseBuilder.jsx (-832 lines). Wire new
modules into the hooks/lib barrels + READMEs.
