+++ .changelog/NEXT.md
@@ -100,3 +100,4 @@
 - **[issue-2375] Split the ~1,380-line `server/routes/apps.js` into domain-focused sub-routers.** The single apps router (24+ endpoints spanning CRUD/status enrichment, PM2 lifecycle, config repair, Xcode provisioning, icon serving, task-type overrides, document git-commits, editor/folder launch, and agent history) is now assembled from nine mountable sub-routers under `server/routes/apps/` (`crud`, `lifecycle`, `viteTls`, `xcode`, `icons`, `taskTypes`, `launch`, `documents`, `agents`), mirroring the pipeline sub-router pattern. Shared plumbing (`loadApp` middleware, `pathExists`, and the port/overall-status derivation helpers previously duplicated between the list and detail endpoints) lives in one `apps/shared.js`, and the ~130-line ecosystem-config port-rewrite orchestration moved out of the PUT handler into a `server/services/appPortConfig.js` service. The `/api/apps` route contract is byte-for-byte identical (all 35 path+method pairs verified) and the oversized route test file is split along the same domain boundaries. No behavior change.
 
 - **[issue-2389] The Sharing page's per-bucket detail tab (Inbox / Activity / Settings) is now URL-derived and history-safe.** The sub-tab was held in component-local `useState('inbox')`, so it wasn't deep-linkable and browser back/forward could disagree with the visible tab. It now derives from a validated `?tab=` search param via `useDrawerTab('tab', 'inbox', BUCKET_TAB_IDS)` — a direct link opens the named tab, a stale/hand-edited value degrades to Inbox, and switching buckets (a fresh route without the query) resets to Inbox. New render tests cover the direct-link, invalid-fallback, and back/forward cases.
+- **[issue-2374] Universe Builder page decomposed — controller hook + shared/count libs + tab components extracted.** `client/src/pages/UniverseBuilder.jsx` was ~3,700 lines with render-job orchestration, shared constants, prompt-count math, and several full tab/editor components all inline. This slice pulls out the independently-testable pieces without changing behavior or the URL-source-of-truth navigation: a `useRenderJobQueue` controller hook (the per-entry `pendingByEntryId` queue, head-map, enqueue/clear reducer), two pure libs — `universeBuilderShared` (trunk/tab/composite constants + category helpers) and `universeBuilderCounts` (server-mirrored `renderPromptCount`/`scopedPromptCount`/canon-content helpers) — and three tab/editor components moved to `components/universeBuilder/` (`CompositeSheetsEditor`, `BucketChipStrip`, `RenderTab`). Each extraction ships with unit tests; the existing `CategoryEditor`/`TrunkView`/`OtherTab`/`UniverseSelector` test seam is preserved. The page drops ~810 lines. Remaining tab extractions (CategoryEditor, VariationCard, BibleTab, TrunkView, OtherTab) and a fuller page-orchestration hook are deferred.
