+++ client/src/hooks/index.js
@@ -127,6 +127,7 @@ export * from './useDeathClock.js';
 export * from './useGoalDetail.js';
 export * from './usePostSession.js';
 export * from './useRecordMerge.js';
+export * from './useRenderJobQueue.js';
 export * from './useSyncIntegrity.js';
 export * from './useTwinEvaluationSuite.js';
 export * from './useUniverseNav.js';
