+++ client/src/lib/index.js
@@ -12,6 +12,8 @@ export * from './seasonStructure.js';
 export * from './sheetPointers.js';
 export * from './shotContinuity.js';
 export * from './shotGrammar.js';
+export * from './universeBuilderCounts.js';
+export * from './universeBuilderShared.js';
 export * from './universeRunTag.js';
 export * from './universeStylePreset.js';
